Add new action capture-logs.
This action allows capturing logs for either a kubernetes
cluster service or docker services, which rely on docker
compose.

Also, modify the triggers for the CI jobs, to be launched
only when pushed to main or when a relevant file is changed
(this way we avoid blocking the merge due to unrelated changes).